[#153756590] jOOQ should be completely encapsulated in the DAO implementation, and the DAO interface should interact with the POJO entity models only.   * Reduce cost of development. We are already seeing a proliferation of spaghetti, which increasingly hinders development as complexity increases.   * Increase portability of the proprietary business logic we are implementing as software. Establish a discreet line, ensuring that our proprietary business logic is implemented independently of any particular data persistence solution (e.g. jOOQ and MySQL in our current implementation)   * User, UserAuth, UserRole, and Access models and services were the first part of the platform to be built-- and turned out to be a complete mess, given the hindsight of a year of further development.
[#153232299] In the Chain-Links index view, ought to see the instruments chosen for each Link, and be able to click an instrument to navigate to that instrument.
  * [ui/hub-ui] Add Arrangement model, related to Link Choices

Also:
  * [core/dao] Refactor such that jOOQ entities exist only below the DAO layer, and Models are used above it as an API
      - Arrangement
      - Choice
      - LinkChord
      - LinkMeme
      - LinkMessage
  * [hub] "Record" resources renamed to "One" for clarity
  * [core/json] Gson object mapper is used globally for writing output JSON
  * [core/gson] Custom TimestampSerializer includes microseconds, ala "2015-05-10 12:17:02.527142Z"
  * [core/timestamp] UTC timestamp parsing must manually remove trailing "Z" from ISO8601 timestamps

+++ core/src/main/java/io/xj/core/model/choice/Chooser.java
@@ -1,14 +1,13 @@
 // Copyright (c) 2017, XJ Music Inc. (https://xj.io) All Rights Reserved.
 package io.xj.core.model.choice;
 
-import io.xj.core.model.Entity;
-
-import org.jooq.types.ULong;
-
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 
+import io.xj.core.model.Entity;
+
 import javax.annotation.Nullable;
+import java.math.BigInteger;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Comparator;
@@ -28,7 +27,7 @@ import java.util.Objects;
 public class Chooser<T extends Entity> {
   //  final TypeToken<T> type = new TypeToken<T>(getClass()) {};
   private final List<T> entities;
-  private final HashMap<ULong, Double> scores;
+  private final HashMap<BigInteger, Double> scores;
 
   /**
    Constructor instantiates a new inner hash map
@@ -80,7 +79,7 @@ public class Chooser<T extends Entity> {
 
    @param entityId to add
    */
-  public void score(ULong entityId, Double Q) {
+  public void score(BigInteger entityId, Double Q) {
     if (scores.containsKey(entityId))
       scores.put(entityId, scores.get(entityId) + Q);
     else
@@ -101,7 +100,7 @@ public class Chooser<T extends Entity> {
 
    @return all entities
    */
-  public Map<ULong,Double> getScores() {
+  public Map<BigInteger, Double> getScores() {
     return Collections.unmodifiableMap(scores);
   }
